Add: Simplify using std::enable_shared_from_this
Adds a new header file `intrusive.h` that simplifies using
`std::enable_shared_from_this` with most C++ types. The header
file provides a `make_intrusive` function that creates a new
type that's derived from the input type and
`std::enable_shared_from_this`, and returns a shared pointer to
the newly instantiated object.

Has a class template that takes the type to be derived from and
restricts the construction to a Create function. This is to
ensure that the object is always created on the heap and not
by value. The Create function is a static member function that
forwards the arguments to the constructor of the derived type and
returns a shared pointer to the newly created object.

Also includes a new test file `intrusive_tests.cpp` that tests
the new functionality.

Signed-off-by: CLV <git@bgtree.com>