Add template for making numeric types strong
This commit adds a template for making numeric types strongly typed.
This is useful for preventing bugs where a function is called with the
wrong type of number. For example, if a function is expecting an offset
and is passed a count. This can be a common source of bugs. It also
allows more meaningful function overloads.

Signed-off-by: CLV <git@bgtree.com>